Module alloy_rpc_types_beacon::events

Beacon API events support. Support for the Beacon API events

See also ethereum-beacon-API eventstream

Modules

attestation

Module for handling attestation-related functionality and events.

light_client_finality

Module for handling light client finality-related functionality and events.

light_client_optimistic

Module for handling light client optimistic-related functionality and events.

Structs

AttestationEvent

Event for the Attestation topic of the beacon API node event stream.

BlobSidecarEvent

Event for the BlobSidecar topic of the beacon API node event stream.

BlockEvent

Event for the Block topic of the beacon API node event stream.

BlsToExecutionChangeEvent

Event for the BlsToExecutionChange topic of the beacon API node event stream.

BlsToExecutionChangeMessage

Represents a BLS to Execution change message in the beacon chain.

ChainReorgEvent

Event for the ChainReorg topic of the beacon API node event stream.

Contribution

Represents a contribution in the beacon chain.

ContributionAndProofEvent

Event for the ContributionAndProof topic of the beacon API node event stream.

ContributionAndProofMessage

Represents a signed contribution and proof message in the beacon chain.

FinalizedCheckpointEvent

Event for the Deposit topic of the beacon API node event stream.

HeadEvent

Event for the Head topic of the beacon API node event stream.

LightClientFinalityUpdateEvent

Event for the LightClientFinalityUpdate topic of the beacon API node event stream.

LightClientOptimisticUpdateEvent

Event for the LightClientOptimisticUpdate topic of the beacon API node event stream.

PayloadAttributesData

Data of the event that contains the payload attributes

PayloadAttributesEvent

Event for the payload_attributes topic of the beacon API node event stream.

VoluntaryExitEvent

Event for the VoluntaryExit topic of the beacon API node event stream.

VoluntaryExitMessage

Represents a voluntary exit message in the beacon chain.

Enums

BeaconNodeEventTopic

Topic variant for the eventstream API